Move DTYPE_MAP to petals.constants
