+++ benchmarks/benchmark_forward.py
@@ -8,7 +8,7 @@ import torch
 from hivemind.utils.logging import get_logger
 
 from petals import AutoDistributedModel
-from petals.constants import PUBLIC_INITIAL_PEERS
+from petals.constants import DTYPE_MAP, PUBLIC_INITIAL_PEERS
 
 logger = get_logger()
 
@@ -16,11 +16,12 @@ logger = get_logger()
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--model", type=str, default="bigscience/bloom")
-    parser.add_argument("-i", "--initial_peers", type=str, nargs="+", default=PUBLIC_INITIAL_PEERS)
-    parser.add_argument("-p", "--n_processes", type=str, default=1)
+    parser.add_argument("--initial_peers", type=str, nargs="+", default=PUBLIC_INITIAL_PEERS)
+    parser.add_argument("--torch_dtype", type=str, default="bfloat16")
+    parser.add_argument("--n_processes", type=str, default=1)
     parser.add_argument("--seq_len", type=int, default=128)
     parser.add_argument("--n_steps", type=int, default=100)
-    parser.add_argument("-b", "--batch_size", type=int, required=True)
+    parser.add_argument("--batch_size", type=int, required=True)
     parser.add_argument("--warmup_steps", type=int, default=1)
     args = parser.parse_args()
 
@@ -39,7 +40,9 @@ def main():
 @torch.inference_mode()
 def benchmark_forward(process_idx, args):
     model = AutoDistributedModel.from_pretrained(
-        args.model, initial_peers=args.initial_peers, torch_dtype=torch.bfloat16
+        args.model,
+        initial_peers=args.initial_peers,
+        torch_dtype=DTYPE_MAP[args.torch_dtype],
     )
     logger.info(f"Created model: {process_idx=} {model.device=}")
