+++ benchmarks/benchmark_training.py
@@ -9,7 +9,7 @@ import torch
 from hivemind.utils.logging import get_logger
 
 from petals import AutoDistributedModelForCausalLM, AutoDistributedModelForSequenceClassification
-from petals.constants import PUBLIC_INITIAL_PEERS
+from petals.constants import DTYPE_MAP, PUBLIC_INITIAL_PEERS
 
 logger = get_logger()
 
@@ -19,12 +19,13 @@ def main():
     parser.add_argument("--model", type=str, default="bigscience/bloom")
     parser.add_argument("--device", type=str, default="cpu")
     parser.add_argument("--task", type=str, default="cls")
-    parser.add_argument("-i", "--initial_peers", type=str, nargs="+", default=PUBLIC_INITIAL_PEERS)
+    parser.add_argument("--initial_peers", type=str, nargs="+", default=PUBLIC_INITIAL_PEERS)
+    parser.add_argument("--torch_dtype", type=str, default="bfloat16")
     parser.add_argument("--n_processes", type=str, default=1)
     parser.add_argument("--seq_len", type=int, default=128)
     parser.add_argument("--pre_seq_len", type=int, default=16)
     parser.add_argument("--n_steps", type=int, default=10)
-    parser.add_argument("-b", "--batch_size", type=int, required=True)
+    parser.add_argument("--batch_size", type=int, required=True)
     parser.add_argument("--warmup_steps", type=int, default=1)
     args = parser.parse_args()
 
@@ -47,13 +48,18 @@ def benchmark_training(process_idx, args):
         model = AutoDistributedModelForSequenceClassification.from_pretrained(
             args.model,
             initial_peers=args.initial_peers,
+            torch_dtype=DTYPE_MAP[args.torch_dtype],
             tuning_mode="deep_ptune",
             pre_seq_len=args.pre_seq_len,
             num_labels=2,
         )
     elif args.task == "causal_lm":
         model = AutoDistributedModelForCausalLM.from_pretrained(
-            args.model, initial_peers=args.initial_peers, tuning_mode="deep_ptune", pre_seq_len=args.pre_seq_len
+            args.model,
+            initial_peers=args.initial_peers,
+            torch_dtype=DTYPE_MAP[args.torch_dtype],
+            tuning_mode="deep_ptune",
+            pre_seq_len=args.pre_seq_len,
         )
     model = model.to(args.device)
     opt = torch.optim.Adam(model.parameters())
